refactor(state): move state classes to state namespace
US-32402
