+++ src/States/State.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Interface of state entity
+ *
+ * @author Denis Ptushko <d.ptushko@artox.com>
+ */
+
+namespace ArtoxLab\Entities\States;
+
+interface State
+{
+
+}
